feat(#147): rename tests table to boots table
- rename the old tests table to boots table
- use the tests table that was not being used and were on tabs/tests
- remove everything related to tests that will not be used by boots table
