+++ dashboard/src/components/Table/TestsTable.tsx
@@ -1,134 +1,128 @@
-import { useCallback, useMemo, useState } from 'react';
+import { useCallback, useMemo } from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 
 import { useNavigate, useSearch } from '@tanstack/react-router';
 
-import { MdChevronRight } from 'react-icons/md';
-
 import BaseTable from '@/components/Table/BaseTable';
 import { TableInfo } from '@/components/Table/TableInfo';
 import { TableCell, TableRow } from '@/components/ui/table';
+import ColoredCircle from '@/components/ColoredCircle/ColoredCircle';
 import { usePagination } from '@/hooks/usePagination';
+import { formatDate } from '@/utils/utils';
+import { MessagesKey } from '@/locales/messages';
 
-import { useTestsByTreeAndCommitHash } from '@/api/TreeDetails';
+import { useTreeTest } from '@/api/TreeDetails';
 
-import {
-  TestsTableFilter,
-  possibleTestsTableFilter,
-} from '@/types/tree/TreeDetails';
-
-import TableStatusFilter from './TableStatusFilter';
-
-const headerLabelIds: string[] = [
-  'Path',
-  'Status',
-  'Start time',
-  'Duration',
-  '', //extra one to add the chevron icon
+const headerLabelIds: MessagesKey[] = [
+  'global.origins',
+  'global.name',
+  'buildDetails.testResults',
+  'buildDetails.startTime',
 ];
+
+const testCellProps = [
+  {
+    name: 'pass_tests',
+    colorClass: 'bg-lightGreen',
+    tooltipLabelId: 'global.pass',
+  },
+  {
+    name: 'done_tests',
+    colorClass: 'bg-green',
+    tooltipLabelId: 'global.done',
+  },
+  {
+    name: 'fail_tests',
+    colorClass: 'bg-lightRed',
+    tooltipLabelId: 'global.failed',
+  },
+  {
+    name: 'error_tests',
+    colorClass: 'bg-red',
+    tooltipLabelId: 'global.error',
+  },
+  {
+    name: 'miss_tests',
+    colorClass: 'bg-yellow',
+    tooltipLabelId: 'global.missed',
+  },
+  {
+    name: 'skip_tests',
+    colorClass: 'bg-darkGray',
+    tooltipLabelId: 'global.skiped',
+  },
+  {
+    name: 'total_tests',
+    colorClass: 'bg-darkGray2',
+    tooltipLabelId: 'global.total',
+  },
+] as const;
 
 const ITEMS_PER_PAGE = 10;
 
 export interface ITestsTable {
   treeId: string;
-  isBootTable?: boolean;
+  patchset?: string;
 }
 
-const TestsTable = ({
-  treeId,
-  isBootTable = false,
-}: ITestsTable): JSX.Element => {
-  const navigate = useNavigate({ from: '/tree/$treeId' });
+const TestsTable = ({ treeId, patchset = '' }: ITestsTable): JSX.Element => {
   const intl = useIntl();
-  const { origin, treeInfo, diffFilter } = useSearch({
-    from: '/tree/$treeId/',
-  });
-  const { tableFilter, currentTreeDetailsTab } = useSearch({
-    from: '/tree/$treeId/',
-  });
-  const { data, error, isLoading } = useTestsByTreeAndCommitHash(
+  const { testPath, treeInfo, origin } = useSearch({ from: '/tree/$treeId/' });
+  const navigate = useNavigate({ from: '/tree/$treeId' });
+  const { data, error } = useTreeTest(
     treeId,
-    isBootTable,
+    patchset,
+    testPath,
+    treeInfo.gitBranch ?? '',
+    treeInfo.gitUrl ?? '',
     origin,
-    treeInfo.gitUrl,
-    treeInfo.gitBranch,
   );
+  const data_len = data?.length || 0;
+  const { startIndex, endIndex, onClickGoForward, onClickGoBack } =
+    usePagination(data_len, ITEMS_PER_PAGE);
 
   const onClickName = useCallback(
-    (id: string) => {
-      navigate({
-        to: '/tree/$treeId/test/$testId',
-        params: {
-          treeId,
-          testId: id,
-        },
-        search: s => s,
-      });
+    (e: React.MouseEvent<HTMLTableCellElement>) => {
+      if (e.target instanceof HTMLTableCellElement) {
+        const newTestPath = e.target.innerText;
+        navigate({
+          search: previousSearch => ({
+            ...previousSearch,
+            testPath: newTestPath,
+          }),
+        });
+      }
     },
-    [navigate, treeId],
+    [navigate],
   );
 
-  const isTestsTab = currentTreeDetailsTab === 'treeDetails.tests';
-  const isBootsTab = currentTreeDetailsTab === 'treeDetails.boots';
-
-  const filteredData = useMemo(() => {
-    const filterToApply = isTestsTab
-      ? tableFilter.testsTable
-      : isBootsTab
-        ? tableFilter.bootsTable
-        : '';
-    if (filterToApply === 'all') {
-      return data?.tests;
-    }
-    return data?.tests.filter(
-      test => test.status.toLowerCase() === filterToApply,
-    );
-  }, [data?.tests, tableFilter, isBootsTab, isTestsTab]);
-
-  const data_len = filteredData?.length || 0;
-
-  const { startIndex, endIndex, onClickGoForward, onClickGoBack } =
-    usePagination(
-      data_len,
-      ITEMS_PER_PAGE,
-      tableFilter.bootsTable,
-      tableFilter.testsTable,
-      tableFilter.buildsTable,
-      diffFilter,
-    );
-
-  const [bootsSelectedFilter, setBootsSelectedFilter] =
-    useState<TestsTableFilter>(tableFilter.bootsTable);
-  const [testsSelectedFilter, setTestsSelectedFilter] =
-    useState<TestsTableFilter>(tableFilter.testsTable);
-
   const headers = useMemo(() => {
-    return headerLabelIds.map(labelId => <p key={labelId}>{labelId}</p>);
+    return headerLabelIds.map(labelId => (
+      <FormattedMessage key={labelId} id={labelId} />
+    ));
   }, []);
 
   const rows = useMemo(() => {
     if (!data || error) return <></>;
 
-    if (filteredData?.length === 0) {
-      return (
-        <div className="flex h-8 items-center px-4">
-          <FormattedMessage id="global.noResults" />
-        </div>
-      );
-    }
-
-    return filteredData?.slice(startIndex, endIndex).map(test => (
-      <TableRow onClick={() => onClickName(test.id)} key={test.id}>
-        <TableCell>{test.path}</TableCell>
-        <TableCell>{test.status}</TableCell>
-        <TableCell>{test.startTime ?? '-'}</TableCell>
-        <TableCell>{test.duration ?? '-'}</TableCell>
-        <TableCell>
-          <MdChevronRight />
+    return data.slice(startIndex, endIndex).map(test => (
+      <TableRow key={test.current_path}>
+        <TableCell>{test.origins.join(', ')}</TableCell>
+        <TableCell onClick={onClickName}>{test.current_path}</TableCell>
+        <TableCell className="flex flex-row gap-1">
+          {testCellProps.map(props => (
+            <ColoredCircle
+              key={test[props.name]}
+              tooltipText={intl.formatMessage({ id: props.tooltipLabelId })}
+              quantity={test[props.name]}
+              backgroundClassName={props.colorClass}
+            />
+          ))}
         </TableCell>
+        <TableCell>{formatDate(test.start_time ?? '')}</TableCell>
       </TableRow>
     ));
-  }, [filteredData, data, error, startIndex, endIndex, onClickName]);
+  }, [data, error, intl, onClickName, startIndex, endIndex]);
 
   const tableInfoElement = (
     <div className="flex flex-col items-end">
@@ -143,92 +137,9 @@ const TestsTable = ({
     </div>
   );
 
-  const onClickFilter = (filter: TestsTableFilter): void => {
-    isBootsTab
-      ? setBootsSelectedFilter(filter)
-      : setTestsSelectedFilter(filter);
-    navigate({
-      search: previousParams => {
-        return {
-          ...previousParams,
-          tableFilter: {
-            testsTable: isTestsTab
-              ? filter
-              : previousParams.tableFilter.testsTable,
-            buildsTable: previousParams.tableFilter.buildsTable,
-            bootsTable: isBootsTab
-              ? filter
-              : previousParams.tableFilter.bootsTable,
-          },
-        };
-      },
-    });
-  };
-
-  const checkIfFilterIsSelected = useCallback(
-    (filter: TestsTableFilter): boolean => {
-      return (
-        (isBootsTab && bootsSelectedFilter === filter) ||
-        (isTestsTab && testsSelectedFilter === filter)
-      );
-    },
-    [isBootsTab, bootsSelectedFilter, isTestsTab, testsSelectedFilter],
-  );
-
-  const filters = useMemo(
-    () => [
-      {
-        label: intl.formatMessage({ id: 'global.all' }),
-        value: possibleTestsTableFilter[0],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[0]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.pass' }),
-        value: possibleTestsTableFilter[5],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[5]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.fail' }),
-        value: possibleTestsTableFilter[3],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[3]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.skip' }),
-        value: possibleTestsTableFilter[6],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[6]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.done' }),
-        value: possibleTestsTableFilter[1],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[1]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.error' }),
-        value: possibleTestsTableFilter[2],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[2]),
-      },
-      {
-        label: intl.formatMessage({ id: 'testStatus.miss' }),
-        value: possibleTestsTableFilter[4],
-        isSelected: checkIfFilterIsSelected(possibleTestsTableFilter[4]),
-      },
-    ],
-    [intl, checkIfFilterIsSelected],
-  );
-
-  if (error) return <FormattedMessage id="global.error" />;
-
-  if (isLoading) return <FormattedMessage id="global.loading" />;
-
   return (
     <div className="flex flex-col gap-6">
-      <div className="flex flex-row justify-between">
-        <TableStatusFilter
-          filters={filters}
-          onClickTest={(filter: TestsTableFilter) => onClickFilter(filter)}
-        />
-        {tableInfoElement}
-      </div>
+      {tableInfoElement}
       <BaseTable headers={headers}>{rows}</BaseTable>
       {tableInfoElement}
     </div>
